requirements.txt
pyserial
pathfinding
pcb-tools-extension
python-telegram-bot
python-dotenv
requests
google-generativeai
chromadb
pypdf
duckduckgo-search
SpeechRecognition
pydub
gTTS
docker
psutil
autopep8
Pillow
beautifulsoup4
PyYAML
pytest